separate out code and refactor
