--- MGSwipeCards/Classes/MGCardStackView/MGCardStackViewDataSource.swift
@@ -1,12 +1,0 @@
-//
-//  MGCardStackViewDataSource.swift
-//  MGSwipeCards
-//
-//  Created by Mac Gallagher on 5/31/18.
-//  Copyright © 2018 Mac Gallagher. All rights reserved.
-//
-
-public protocol MGCardStackViewDataSource {
-    func numberOfCards(in cardStack: MGCardStackView) -> Int
-    func cardStack(_ cardStack: MGCardStackView, cardForIndexAt index: Int) -> MGSwipeCard
-}
