+++ MGSwipeCards/Classes/MGSwipeCard/CardAnimationOptions.swift
@@ -1,0 +1,26 @@
+//
+//  CardOptions.swift
+//  MGSwipeCards
+//
+//  Created by Mac Gallagher on 11/2/18.
+//
+
+open class CardAnimationOptions: NSObject {
+    public static var defaultOptions = CardAnimationOptions()
+    
+    /// The duration of the animated swipe translation. Measured in seconds. Defaults to 0.8.
+    open var cardSwipeAnimationDuration: TimeInterval = 0.8
+    
+    /// The duration of the fade animation applied to the overlays before the animated swipe translation, and after the reverse swipe translation.
+    open var overlayFadeAnimationDuration: TimeInterval = 0.15
+    
+    /// The duration of the animated reverse swipe translation. Measured in seconds. Defaults to 0.2.
+    open var reverseSwipeAnimationDuration: TimeInterval = 0.2
+    
+    ///The effective bounciness of the spring animation upon a cancelled swipe. Higher values increase spring movement range resulting in more oscillations and springiness. Defined as a value in the range [0, 20]. Defaults to 12.
+    open var resetAnimationSpringBounciness: CGFloat = 12.0
+    
+    /// The effective speed of the spring animation upon a cancelled swipe. Higher values increase the dampening power of the spring. Defined as a value in the range [0, 20]. Defaults to 20.
+    open var resetAnimationSpringSpeed: CGFloat = 20.0
+    
+}
